+++ src/main/java/com/pbs/ams/web/mappers/UpmsApiMapper.java
@@ -13,9 +13,9 @@ import java.util.List;
 public interface UpmsApiMapper {
 
 	// 根据用户id获取所拥有的权限
-	List<UpmsPermission> selectUpmsPermissionByUpmsUserId(Integer upmsUserId);
+	List<UpmsPermission> selectUpmsPermissionByUpmsUserId(Long upmsUserId);
 
 	// 根据用户id获取所属的角色
-	List<UpmsRole> selectUpmsRoleByUpmsUserId(Integer upmsUserId);
+	List<UpmsRole> selectUpmsRoleByUpmsUserId(Long upmsUserId);
 	
 }
