+++ src/main/java/com/pbs/ams/web/mappers/UpmsRoleMapper.java
@@ -10,7 +10,7 @@ public interface UpmsRoleMapper {
 
     int deleteByExample(UpmsRoleExample example);
 
-    int deleteByPrimaryKey(Integer roleId);
+    int deleteByPrimaryKey(Long roleId);
 
     int insert(UpmsRole record);
 
@@ -18,7 +18,7 @@ public interface UpmsRoleMapper {
 
     List<UpmsRole> selectByExample(UpmsRoleExample example);
 
-    UpmsRole selectByPrimaryKey(Integer roleId);
+    UpmsRole selectByPrimaryKey(Long roleId);
 
     int updateByExampleSelective(@Param("record") UpmsRole record, @Param("example") UpmsRoleExample example);
